1) while adding event through fee collection, start date and end date is not being inserted in event table. 2) "Fees Paid" field was not being updated even when all pending fees were paid by students. 3) User Access module need to be updated to have Privilege incorporated in the implementation. 4) In Store Module, there is an error when product etails are being updated. 5) In Product Details edit screen, Product cost per unit has to be made decimal.

+++ SFSAcademy/SFSAcademy/STORE_PURCHAGE.cs
@@ -17,7 +17,7 @@ namespace SFSAcademy
         public int ID { get; set; }
         public int PRODUCT_ID { get; set; }
         public Nullable<int> UNIT_SOLD { get; set; }
-        public Nullable<int> SOLD_PRICE { get; set; }
+        public Nullable<decimal> SOLD_PRICE { get; set; }
         public string SOLD_BY { get; set; }
         public Nullable<System.DateTime> SOLD_ON { get; set; }
         public string STUDENT_NAME { get; set; }
